circle_mini_project

+++ calculator_mini_project/main.py
@@ -1,5 +1,5 @@
-first_number = float(input("Enter the first number: "))
-second_number = float(input("Enter the second number: "))
+first_number = int(input("Enter the first number: "))
+second_number = int(input("Enter the second number: "))
 sum =first_number+second_number
 sub=first_number-second_number
 mul=first_number*second_number
